- Implemented GetInstancesFromQuery - Moved shareable code from InstanceRepository to InstanceQueryHelper - Added index for partyId
