+++ src/Storage/Migration/v0.00/01-setup-tables.sql
@@ -29,4 +29,5 @@ CREATE TABLE IF NOT EXISTS storage.instanceEvents
 )
 TABLESPACE pg_default;
 
-CREATE INDEX dataElements_instanceInternalId ON storage.dataElements(instanceInternalId);
+CREATE INDEX IF NOT EXISTS dataElements_instanceInternalId ON storage.dataElements(instanceInternalId);
+CREATE INDEX IF NOT EXISTS instances_partyId ON storage.instances(partyId);
