+++ src/Storage/Repository/InstanceRepository.cs
@@ -100,7 +100,7 @@ namespace Altinn.Platform.Storage.Repository
 
                 try
                 {
-                    queryBuilder = BuildQueryFromParameters(queryParams, queryBuilder, options);
+                    queryBuilder = InstanceQueryHelper.BuildQueryFromParameters(queryParams, queryBuilder, options);
                 }
                 catch (Exception e)
                 {
@@ -143,373 +143,6 @@ namespace Altinn.Platform.Storage.Repository
             }
 
             return queryResponse;
-        }
-
-        private static IQueryable<Instance> BuildQueryFromParameters(Dictionary<string, StringValues> queryParams, IQueryable<Instance> queryBuilder, QueryRequestOptions options)
-        {
-            foreach (KeyValuePair<string, StringValues> param in queryParams)
-            {
-                string queryParameter = param.Key;
-                StringValues queryValues = param.Value;
-
-                if (queryParameter.Equals("appId"))
-                {
-                    queryBuilder = queryBuilder.Where(i => queryValues.Contains(i.AppId));
-                    continue;
-                }
-
-                if (queryParameter.Equals("instanceOwner.partyId"))
-                {
-                    if (queryValues.Count == 1)
-                    {
-                        var partyId = queryValues.First();
-                        options.PartitionKey = new PartitionKey(partyId);
-                        queryBuilder = queryBuilder.Where(i => partyId == i.InstanceOwner.PartyId);
-                    }
-                    else
-                    {
-                        queryBuilder = queryBuilder.Where(i => queryValues.Contains(i.InstanceOwner.PartyId));
-                    }
-
-                    continue;
-                }
-
-                foreach (string queryValue in queryValues)
-                {
-                    switch (queryParameter)
-                    {
-                        case "size":
-                        case "continuationToken":
-                            // handled outside this method, it is a valid parameter.
-                            break;
-                        case "org":
-                            queryBuilder = queryBuilder.Where(i => i.Org == queryValue);
-                            break;
-                        case "lastChanged":
-                            queryBuilder = QueryBuilderForLastChangedDateTime(queryBuilder, queryValue);
-                            break;
-
-                        case "dueBefore":
-                            queryBuilder = QueryBuilderForDueBefore(queryBuilder, queryValue);
-                            break;
-
-                        case "visibleAfter":
-                            queryBuilder = QueryBuilderForVisibleAfter(queryBuilder, queryValue);
-                            break;
-
-                        case "created":
-                            queryBuilder = QueryBuilderForCreated(queryBuilder, queryValue);
-                            break;
-
-                        case "process.currentTask":
-                            queryBuilder = queryBuilder.Where(i => i.Process.CurrentTask.ElementId == queryValue);
-                            break;
-
-                        case "process.isComplete":
-                            bool isComplete = bool.Parse(queryValue);
-                            if (isComplete)
-                            {
-                                queryBuilder = queryBuilder.Where(i => i.Process.Ended != null);
-                            }
-                            else
-                            {
-                                queryBuilder = queryBuilder.Where(i => i.Process.CurrentTask != null);
-                            }
-
-                            break;
-
-                        case "process.ended":
-                            queryBuilder = QueryBuilderForEnded(queryBuilder, queryValue);
-                            break;
-
-                        case "excludeConfirmedBy":
-                            queryBuilder = QueryBuilderExcludeConfirmedBy(queryBuilder, queryValue);
-                            break;
-                        case "language":
-                            break;
-                        case "status.isArchived":
-                            bool isArchived = bool.Parse(queryValue);
-                            queryBuilder = queryBuilder.Where(i => i.Status.IsArchived == isArchived);
-
-                            break;
-                        case "status.isSoftDeleted":
-                            bool isSoftDeleted = bool.Parse(queryValue);
-                            queryBuilder = queryBuilder.Where(i => i.Status.IsSoftDeleted == isSoftDeleted);
-
-                            break;
-                        case "status.isHardDeleted":
-                            bool isHardDeleted = bool.Parse(queryValue);
-                            queryBuilder = queryBuilder.Where(i => i.Status.IsHardDeleted == isHardDeleted);
-
-                            break;
-                        case "status.isArchivedOrSoftDeleted":
-                            if (bool.Parse(queryValue))
-                            {
-                                queryBuilder = queryBuilder.Where(i => i.Status.IsArchived || i.Status.IsSoftDeleted);
-                            }
-
-                            break;
-                        case "status.isActiveOrSoftDeleted":
-                            if (bool.Parse(queryValue))
-                            {
-                                queryBuilder = queryBuilder.Where(i => !i.Status.IsArchived || i.Status.IsSoftDeleted);
-                            }
-
-                            break;
-                        case "sortBy":
-                            queryBuilder = QueryBuilderForSortBy(queryBuilder, queryValue);
-
-                            break;
-                        case "archiveReference":
-                            queryBuilder = queryBuilder.Where(i => i.Id.EndsWith(queryValue.ToLower()));
-
-                            break;
-                        default:
-                            throw new ArgumentException($"Unknown query parameter: {queryParameter}");
-                    }
-                }
-            }
-
-            return queryBuilder;
-        }
-
-        // Limitations in queryBuilder.Where interface forces me to duplicate the datetime methods
-        private static IQueryable<Instance> QueryBuilderForDueBefore(IQueryable<Instance> queryBuilder, string queryValue)
-        {
-            DateTime dateValue;
-
-            if (queryValue.StartsWith("gt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.DueBefore > dateValue);
-            }
-
-            if (queryValue.StartsWith("gte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.DueBefore >= dateValue);
-            }
-
-            if (queryValue.StartsWith("lt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.DueBefore < dateValue);
-            }
-
-            if (queryValue.StartsWith("lte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.DueBefore <= dateValue);
-            }
-
-            if (queryValue.StartsWith("eq:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.DueBefore == dateValue);
-            }
-
-            dateValue = ParseDateTimeIntoUtc(queryValue);
-            return queryBuilder.Where(i => i.DueBefore == dateValue);
-        }
-
-        // Limitations in queryBuilder.Where interface forces me to duplicate the datetime methods
-        private static IQueryable<Instance> QueryBuilderForLastChangedDateTime(IQueryable<Instance> queryBuilder, string queryValue)
-        {
-            DateTime dateValue;
-
-            if (queryValue.StartsWith("gt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.LastChanged > dateValue);
-            }
-
-            if (queryValue.StartsWith("gte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.LastChanged >= dateValue);
-            }
-
-            if (queryValue.StartsWith("lt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.LastChanged < dateValue);
-            }
-
-            if (queryValue.StartsWith("lte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.LastChanged <= dateValue);
-            }
-
-            if (queryValue.StartsWith("eq:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.LastChanged == dateValue);
-            }
-
-            dateValue = ParseDateTimeIntoUtc(queryValue);
-            return queryBuilder.Where(i => i.LastChanged == dateValue);
-        }
-
-        // Limitations in queryBuilder.Where interface forces me to duplicate the datetime methods
-        private static IQueryable<Instance> QueryBuilderForEnded(IQueryable<Instance> queryBuilder, string queryValue)
-        {
-            DateTime dateValue;
-
-            if (queryValue.StartsWith("gt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.Process.Ended > dateValue);
-            }
-
-            if (queryValue.StartsWith("gte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.Process.Ended >= dateValue);
-            }
-
-            if (queryValue.StartsWith("lt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.Process.Ended < dateValue);
-            }
-
-            if (queryValue.StartsWith("lte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.Process.Ended <= dateValue);
-            }
-
-            if (queryValue.StartsWith("eq:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.Process.Ended == dateValue);
-            }
-
-            dateValue = ParseDateTimeIntoUtc(queryValue);
-            return queryBuilder.Where(i => i.Process.Ended == dateValue);
-        }
-
-        private static IQueryable<Instance> QueryBuilderExcludeConfirmedBy(IQueryable<Instance> queryBuilder, string queryValue)
-        {
-            return queryBuilder.Where(i =>
-
-                // A slightly more readable variant would be to use All( != ), but All() isn't supported.
-                !i.CompleteConfirmations.Any(cc => cc.StakeholderId == queryValue));
-        }
-
-        private static IQueryable<Instance> QueryBuilderForSortBy(IQueryable<Instance> queryBuilder, string queryValue)
-        {
-            string[] value = queryValue.Split(':');
-            string direction = value[0].ToLower();
-            string property = value[1];
-
-            if (!direction.Equals("desc") && !direction.Equals("asc"))
-            {
-                throw new ArgumentException($"Invalid direction for sorting: {direction}");
-            }
-
-            switch (property)
-            {
-                case "lastChanged":
-                    if (direction.Equals("desc"))
-                    {
-                        queryBuilder = queryBuilder.OrderByDescending(i => i.LastChanged);
-                    }
-                    else
-                    {
-                        queryBuilder = queryBuilder.OrderBy(i => i.LastChanged);
-                    }
-
-                    break;
-                default:
-                    throw new ArgumentException($"Cannot sort on property: {property}");
-            }
-
-            return queryBuilder;
-        }
-
-        // Limitations in queryBuilder.Where interface forces me to duplicate the datetime methods
-        private static IQueryable<Instance> QueryBuilderForCreated(IQueryable<Instance> queryBuilder, string queryValue)
-        {
-            DateTime dateValue;
-
-            if (queryValue.StartsWith("gt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.Created > dateValue);
-            }
-
-            if (queryValue.StartsWith("gte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.Created >= dateValue);
-            }
-
-            if (queryValue.StartsWith("lt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.Created < dateValue);
-            }
-
-            if (queryValue.StartsWith("lte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.Created <= dateValue);
-            }
-
-            if (queryValue.StartsWith("eq:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.Created == dateValue);
-            }
-
-            dateValue = ParseDateTimeIntoUtc(queryValue);
-            return queryBuilder.Where(i => i.Created == dateValue);
-        }
-
-        // Limitations in queryBuilder.Where interface forces me to duplicate the datetime methods
-        private static IQueryable<Instance> QueryBuilderForVisibleAfter(IQueryable<Instance> queryBuilder, string queryValue)
-        {
-            DateTime dateValue;
-
-            if (queryValue.StartsWith("gt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.VisibleAfter > dateValue);
-            }
-
-            if (queryValue.StartsWith("gte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.VisibleAfter >= dateValue);
-            }
-
-            if (queryValue.StartsWith("lt:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.VisibleAfter < dateValue);
-            }
-
-            if (queryValue.StartsWith("lte:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(4));
-                return queryBuilder.Where(i => i.VisibleAfter <= dateValue);
-            }
-
-            if (queryValue.StartsWith("eq:"))
-            {
-                dateValue = ParseDateTimeIntoUtc(queryValue.Substring(3));
-                return queryBuilder.Where(i => i.VisibleAfter == dateValue);
-            }
-
-            dateValue = ParseDateTimeIntoUtc(queryValue);
-            return queryBuilder.Where(i => i.VisibleAfter == dateValue);
-        }
-
-        private static DateTime ParseDateTimeIntoUtc(string queryValue)
-        {
-            return DateTimeHelper.ParseAndConvertToUniversalTime(queryValue);
         }
 
         /// <inheritdoc/>
